Simplified design: no mutation after allocation.
Allowing for a `BuilderRef` in a builder context and freezing to an
immutable `Ref` outside the context *almost* works, except it provides a
loophole for unsafety: the user may assign the `BuilderRef` immediately
to a `Ref` within the builder context, then dereference that `Ref`, and
obtain an immutable borrow to the object while a mutable borrow is also
outstanding.

A possible solution to this problem is to defer all ref assignments to
the builder teardown (`Drop` implementation), but this is IMHO
unnecessarily complex, and may produce surprising behavior: (i) the user may
expect a ref to be set and be able to follow it, and (ii) it's somewhat
complicated to track whether an assignment has been enqueued to a ref
already. So for now, this should suffice.